main.py
#!/usr/bin/env python3
"""
Traffic Camera PoC - Main Entry Point

AI-powered traffic monitoring system for residential areas.
Monitors vehicle speed and traffic volume, generates reports for authorities.
"""

import argparse
import logging
import sys
from pathlib import Path

import yaml


def setup_logging(config: dict):
    """Setup logging configuration"""
    log_config = config.get('logging', {})
    level = getattr(logging, log_config.get('level', 'INFO'))

    handlers = []

    # File handler
    log_file = log_config.get('file', './logs/traffic_monitor.log')
    Path(log_file).parent.mkdir(parents=True, exist_ok=True)

    file_handler = logging.FileHandler(log_file)
    file_handler.setLevel(level)
    handlers.append(file_handler)

    # Console handler
    if log_config.get('console', True):
        console_handler = logging.StreamHandler()
        console_handler.setLevel(level)
        handlers.append(console_handler)

    # Configure logging
    logging.basicConfig(
        level=level,
        format='%(asctime)s - %(name)s - %(levelname)s - %(message)s',
        handlers=handlers
    )


def load_config(config_path: str) -> dict:
    """Load configuration from YAML file"""
    try:
        with open(config_path, 'r') as f:
            config = yaml.safe_load(f)
        return config
    except FileNotFoundError:
        print(f"Error: Configuration file not found: {config_path}")
        sys.exit(1)
    except yaml.YAMLError as e:
        print(f"Error: Invalid YAML in configuration file: {e}")
        sys.exit(1)


def main():
    """Main entry point"""
    parser = argparse.ArgumentParser(
        description='Traffic Camera PoC - AI-powered traffic monitoring'
    )
    parser.add_argument(
        '--config',
        type=str,
        default='config.yaml',
        help='Path to configuration file (default: config.yaml)'
    )
    parser.add_argument(
        '--calibrate',
        action='store_true',
        help='Run calibration tool'
    )
    parser.add_argument(
        '--test',
        action='store_true',
        help='Test camera connection and display stream'
    )

    args = parser.parse_args()

    # Load configuration
    config = load_config(args.config)

    # Setup logging
    setup_logging(config)
    logger = logging.getLogger(__name__)

    logger.info("=" * 60)
    logger.info("Traffic Camera PoC Starting")
    logger.info("=" * 60)

    # Calibration mode
    if args.calibrate:
        logger.info("Starting calibration mode...")
        # TODO: Import and run calibration tool
        print("Calibration mode - Not implemented yet")
        return

    # Test mode
    if args.test:
        logger.info("Testing camera connection...")
        # TODO: Import and run camera test
        print("Test mode - Not implemented yet")
        return

    # Normal operation
    logger.info("Starting traffic monitoring...")
    logger.info(f"Configuration loaded from: {args.config}")

    try:
        # TODO: Initialize modules
        # 1. Camera capture
        # 2. Detection
        # 3. Tracking
        # 4. Speed estimation
        # 5. Storage

        logger.info("System initialized successfully")

        # TODO: Start main processing loop
        logger.info("Processing started. Press Ctrl+C to stop.")

        # Placeholder for main loop
        import time
        while True:
            time.sleep(1)

    except KeyboardInterrupt:
        logger.info("Shutting down gracefully...")
    except Exception as e:
        logger.error(f"Fatal error: {e}", exc_info=True)
        sys.exit(1)
    finally:
        logger.info("Traffic Camera PoC stopped")
        logger.info("=" * 60)


if __name__ == "__main__":
    main()
